feat(mirror): add a config file with a tx_send_interval option (#11098)
This option will allow setting the delay between sending each mainnet
block's worth of transactions. If it is not present, we will keep the
old behavior of waiting about as long as blocks in the target chain are
taking to be produced, and if it is present, we'll wait for however long
is indicated between sending each batch. The load that this results in
depends on the number of transactions in each mainnet block we're
sending, but in the future it would be good to implement another option
that allows targeting a specific number of transactions per second, but
this is easier to implement for now.