feat(mocknet): add subcommand to mirror.py to run commands on all mocknet hosts (#11160)
This PR introduces `run-cmd` subcommand that allow the user to issue
commands to all the mocknet hosts.

It can target `--all` the host, only `--nodes` or only `--traffic`
hosts.
It can also `--filter` the selected nodes with a regex.
It prints out the output of the command on every host. ` > /dev/null` if
you don't want it.