refactor: generalize state trie queue (#11165)
In #11159 we added a helper for the delayed receipts queue. To add more
queues as simply as possible, we need to generalize the implementation a
bit more.

First: Move code that is independent of the queue content to a new trait
`TrieQueue`. Trie keys and access to fields is done through trait
methods that each specific queue impl needs to define.

Second: Introduce `struct TrieQueueIndices` for a general way of storing
queue indices. We keep the type `DelayedReceiptIndices` for backwards
compatibility of dependencies to the primitives crate but provide
conversion from and to it.

Third: Change the internals of `ReceiptIterator` to make use of these
generalizations. We can avoid boxing a weird closure this way and make
the addition of new queue implementations much easier.