add congestion info fields to chunk header (#11128)
This is a part of NEP-539.

This PR alone only adds the fields, without actually making use of them.
But it adds all the necessary data for the congestion control strategy
to be able to function. This makes this PR by far the largest.

Splitting the feature into multiple PRs is supposed to make reviewing it
easier.

---------

Co-authored-by: wacban <wac.banasik@gmail.com>