[stateless_validation] client_sender for state_witness_actor to send decoded state witness from parts (#11162)
This PR adds an actix message from state witness actor to client actor
to send the encoded state witness.

The state witness is generated from the partial_witnesses got from the
other chunk validators. We generate and send the state witness to client
as soon as we have enough parts for reed solomon decoding.