Save observed instances of ChunkStateWitness to the database for later analysis (#11137)
Save all of the observed instances of `ChunkStateWitness` to the
database.
They can later be fetched for debugging and analysis.

Sometimes things go wrong with witness validation, and we're not sure
why. In such cases it would be good to take a look at the witness that
failed the validation, but currently there's no way to see the witness,
it disappears after validation. Saving the witnesses in the database
will allow us to inspect the witnesses after it failed validation and
see what exactly is wrong, even after node crash.

Adds a new database column `DBCol::LatestChunkStateWitnesses`, which
keeps a set of latest observed witnesses. Size of this set is limited to
`4GB` and `60*30 (30 min)` of instances. When the limit is hit, the
witness with the lowest height is removed from the set.
We can't really save all witnesses, at 4MB/s that would add up to
345GB/day, so we're forced to garbage collect the old witnesses. Having
only the latest witnesses should be enough for debugging and analysis.

Witness saving could potentially be an attack vector, as we save all
witnesses, even the ones that failed validation. An attacker could spam
the node with thousands of witnesses, which would all be saved to the
database, which could cause denial of service.
Because of that the feature is guarded by a new config option:
`save_latest_witnesses`. By default it's false, so production nodes
won't save anything, and won't be vulnerable to such attacks. It can be
selectively enabled on test/canary nodes when needed.

A new database command is added to access the saved witnesses:
```console
Print observed ChunkStateWitnesses at the given block height (and shard id).
Observed witnesses are only saved when `save_latest_witnesses` is set to true in config.json

Usage: neard database show-latest-witnesses [OPTIONS] --height <HEIGHT>

Options:
      --height <HEIGHT>      Block height (required)
      --shard-id <SHARD_ID>  Shard id (optional)
      --pretty               Pretty-print using the "{:#?}" formatting
      --binary               Print the raw &[u8], can be pasted into rust code
  -h, --help                 Print help
```
The tool allows to print observed witnesses with given height (and
optionally shard id).
Either as a debug print `{:?}/{:#?}`, or as a binary blob that can be
pasted into rust code.

I'm not sure if that's the best way to expose it for debugging, but it
was the easiest one to implement. Maybe it should somehow be integrated
with `debug-ui`? I'm not very familiar with it, idk if it'd make sense.


Fixes: https://github.com/near/nearcore/issues/11110
Similar to: https://github.com/near/nearcore/pull/10599