cleanup(congestion_control) - resolved a few minor TODOs  (#11211)
Removing a few TODOs and updating some others.

* Renamed all TODOs related to resharding to "integration with
resharding" so that it's easier to filter those out.
* Resolved TODO in chunk validation when getting genesis transition
params - reading epoch id directly from header
* Resolved TODOs in ChunkExtra::new - uncommented the asserts. 
* Resolved TODO is validate_prepared_transactions - I checked that the
method that calls this function reads multiple blocks including this
one. It's not convenient to pass the block as an argument but it should
be in the db cache so fetching it again should be really cheap.