fix: Pass release type to binary_release.sh in custom-release GitHub action (#11234)
We added a [new GitHub
action](https://github.com/near/nearcore/actions/workflows/neard_custom_release.yml)
to pass a release type param to the binary_release.sh script, so that we
can compile neard with different features (eg. with stateless
validation). But missed passing the parameter to the script. This change
passes the param value to the script.