feat: Add metrics to measure compiled-contract cache hits/misses for chunk production vs. chunk validation. (#11178)
As part of #11125, we add metrics to count compiled-contract cache
hits/misses.

We want to track the use of compiled-contract cache for apply-chunk
operations. With stateless validation, apply-chunk will be executed for
two different purposes: (1) while updating the shards by chunk producers
as part of tracking shards and (2) while validating the chunk-witness in
stateless validation (they do not necessarily track the shards they
validate). Since the chunk validators are stateless, they will need to
recompile contracts that they see anew unless they have already cached
the previous compilation results.

We want to closely monitor the trend for chunk validators, thus we need
to distinguish the apply-chunk operations executed for chunk producers
vs chunk validators.

In this PR, we introduce a new enum to make this distinction and
propagate it through apply-chunk call stack up to the point VM is
called. Then we use the Metrics structure exported by the VM.run method
to report the metrics.