congestion: only init outgoing buffers when used (#11266)
The previous implementation wrote the outgoing buffer indices to the
trie in `pop_n` even if the buffer was still empty. To prevent this as
showing up as a trie change, I added logic that checks specifically for
empty values.

As it turns out, it is much easier to prevent writing indices back in
the first place, which is also how it's done for the existing delayed
receipts queue to avoid the same issue.