fix(forknet): ignore tracing server and set up state dumper node properly (#11271)
This adds some fields (`can_validate`, `want_state_dump`,
`want_neard_runner`) to `NodeHandle`. If `want_neard_runner` is false,
then we ignore it when running all commands. If `want_state_dump` is
true, we configure its state sync dump config, along with configuring it
and the rest of the nodes with a correct `state_sync.sync` config. If
`can_validate` is false, we ignore it when building the list of
validator nodes.

The nodes will be configured in `node_config.py`, where for now we match
on the names of instances

---------

Co-authored-by: Razvan Barbascu <razvan@near.org>