feat(congestion_control) - handling missing chunks  (#11274)
When there are multiple missing chunks in a row in a shard we want to
consider that shard as congested. That is in order to prevent outgoing
receipts to that shard accumulating and then blowing up the state
witness size.

I decided to not embed the information about missing chunks in the
congestion info but rather I added a new struct called
ExtendedCongestionInfo. The Block now constructs congestion info,
extends it with the information about missing chunks and provides this
new struct to the runtime.

Since from now the congestion level cannot be calculated without the
missing chunks information I added `missing_chunks_count` argument to
all methods that rely on the congestion level. That is to make sure the
users of those structs do not forget about the missing chunks - compiler
will warn them about it. In the runtime the ExtendedCongestionInfo
struct acts as a helper to make it as convenient as it used to be.

The congestion level itself is now a maximum of 4 values - the first
three as before and a new one for missing chunks. In this PR I made it
so that 10 missed chunks in a row would lead to full congestion - that
number is to be adjusted based on data. Other changes can also be
considered such as adding the missing chunks congestion to the max of
the others. I'm open for suggestions here.


It's in draft because I still need to add tests for this.