Don't run shadow chunk validation when syncing blocks (#11293)
Currently we run shadow validation for every new block, even if we're
only syncing up to reach the current tip of the chain. This causes sync
to be extremely slow, as every synced block is applied. Because of that
it's often faster to setup a new node from scratch rather than wait for
a node to sync.

Let's not run shadow validation when syncing, it'll make it possible to
catch up with the rest of the network in a reasonable amount of time.