Introduce priority fee field for transaction and receipt (#11228)
A first step towards https://github.com/near/NEPs/pull/541 by
introducing a priority field in both transaction and receipt. This is
not entirely trivial due to the need to maintain backward compatibility.
This PR accomplishes backward compatibility by leveraging the account id
serialization and implement manual deserialization for the new
transaction and receipt structures. While this PR appears to be quite
large, most of the changes are trivial. The core of the changes are the
serialization/deserialization of transaction and receipt.

While this change introduces the new versions, they are prohibited from
being used in the current protocol until the introduction of the
protocol change that leverages priorities.