test(nayduck) - Added a test for chain recovery after a really slow chunk (#11317)
Added a test where one chunk takes a really long time to apply. This is
to simulate extreme undercharging. Under those circumstances we want the
blockchain to be able to recover - to eventually start processing chunks
in the affected shard again. In the worst case scenario the

This test currently fails on nightly where stateless validation is
enabled. This needs to be properly investigated and resolved. My
suspicion is that the chain cannot get a new chunk in after the slow
chunk. I think every time a chunk producer attempts to produce a new
chunk, it* takes too long and it doesn't make it to the next block. Once
the next block is produced the next chunk producer needs to start from
scratch.

`*` "it" can actually be different things. It could be that just
producing the chunk after the slow one is too slow. Or it may actually
be produced but then it's too slow to validate it - the validators set
changes every block and needs to apply the slow chunk. Or it could be
something else too on the end to end path of producing chunks and
including chunks into blocks.
  
@akhi3030 and @nagisa Can you have a look at the sleep host function and
advise how to properly gate it?

@tayfunelmas and @robin-near Can you have a look at the python test? 

For reference here is a sample output
https://gist.github.com/wacban/7db697121b7fbeb64c392224a05dbbda