fix: Set protocol reward rate to 1/10 for mocknet (#11341)
Currently protocol_reward_rate (configures the reward for the
validators) is not set in genesis config for mocknet. Based on [this
discussion](https://near.zulipchat.com/#narrow/stream/308695-nearone.2Fprivate/topic/Forknet.20Automation/near/439272276),
the code is not falling back to the hardcoded value of 1/10 because
genesis and protocol versions are the same, so the code is attempting to
get it from the genesis config but the value there is 0 by default.

We update the amend-genesis command to set the reward rate to 1/10.

**I am not adding any option/param for the `mirror new-test` command to
enable/disable it, is there a case that we do not want it in mocknet?**