congestion: genesis congestion info (#11362)
A small PR to initialize the congestion info for genesis.

Please check if this is correct. But I think we can just take the
genesis block and read the congestion info from there. I assume it's
usually empty, but if genesis somehow contains delayed or buffered
receipts, I assume it should be included in the chunk header.